docs/trading_ai_system_updated/trading_ai_system/inference_service/requirements.txt
kafka-python>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
torch>=2.0.0
stable-baselines3>=2.2.1
shap>=0.43.0
lime>=0.2.0
python-dotenv>=1.0.0
tenacity>=8.2.3